feat: Add back navigation to Contact page
Introduces a back button to the Contact page, allowing users to easily return to the homepage. Also, refactors the "Start Your Project" button on the Work page to use the default outline variant for consistency. Updates the styling of the authority statement section on the Work page for a cleaner visual presentation.